Prompts and feeback improvements for gameworld
Added prompt that shows when player tries to pickup up something when invertory is full. Increased the time the hit effect is on the screen to make it more noticiable that the player is taking damage. Added affect to enemies that makes them flash when they are in their exploading state.  Updated the trigger on the safehouse so that it matches the house geo more. Lastly lowered the volume on all music in the gameworld so it is not so loud.